+++ tests/unit/ska_tmc_cdm/serialisation/subarray_node/test_configure.py
@@ -26,10 +26,13 @@ from ska_tmc_cdm.messages.subarray_node.configure.csp import (
     BeamsConfiguration,
     CBFConfiguration,
     CommonConfiguration,
+    CorrelationConfiguration,
     CSPConfiguration,
     FSPConfiguration,
     FSPFunctionMode,
     LowCBFConfiguration,
+    MidCBFConfiguration,
+    ProcessingRegionConfiguration,
     StationConfiguration,
     StnBeamConfiguration,
     SubarrayConfiguration,
@@ -102,11 +105,11 @@ NON_COMPLIANCE_MID_CONFIGURE_OBJECT = ConfigureRequest(
         scan_type="science_A",
     ),
     csp=CSPConfiguration(
-        interface="https://schema.skao.int/ska-csp-configure/2.0",
+        interface="https://schema.skao.int/ska-csp-configurescan/2.0",
         subarray=SubarrayConfiguration(subarray_name="science period 23"),
         common=CommonConfiguration(
             config_id="sbi-mvp01-20200325-00001-science_A",
-            frequency_band=ReceiverBand.BAND_5B,
+            frequency_band=ReceiverBand.BAND_2,
             subarray_id=1,
         ),
         pss_config={},
@@ -188,13 +191,13 @@ NON_COMPLIANCE_MID_CONFIGURE_JSON = """
     "receiver_band": "5a"
   },
   "csp": {
-    "interface": "https://schema.skao.int/ska-csp-configure/2.0",
+    "interface": "https://schema.skao.int/ska-csp-configurescan/2.0",
     "subarray": {
       "subarray_name": "science period 23"
     },
     "common": {
       "config_id": "sbi-mvp01-20200325-00001-science_A",
-      "frequency_band": "5b",
+      "frequency_band": "2",
       "subarray_id": 1
     },
     "cbf": {
@@ -988,7 +991,7 @@ VALID_NULL_JSON = (
 
 VALID_NULL_OBJECT = ConfigureRequest(interface=MID_SCHEMA)
 
-VALID_MID_CONFIGURE_JSON = """
+VALID_MID_CONFIGURE_JSON_2_3 = """
 {
   "interface": "https://schema.skao.int/ska-tmc-configure/2.3",
   "transaction_id": "txn-....-00001",
@@ -1005,7 +1008,7 @@ VALID_MID_CONFIGURE_JSON = """
     "receiver_band": "1"
   },
   "csp": {
-    "interface": "https://schema.skao.int/ska-csp-configure/2.0",
+    "interface": "https://schema.skao.int/ska-csp-configurescan/2.0",
     "subarray": {
       "subarray_name": "science period 23"
     },
@@ -1088,6 +1091,65 @@ VALID_MID_CONFIGURE_JSON = """
   }
 }"""
 
+VALID_MID_CONFIGURE_JSON_4_0 = """
+{
+  "interface": "https://schema.skao.int/ska-tmc-configure/4.0",
+  "transaction_id": "txn-....-00001",
+  "pointing": {
+    "target": {
+      "reference_frame": "ICRS",
+      "target_name": "Polaris Australis",
+      "ra": "21:08:47.92",
+      "dec": "-88:57:22.9"
+    },
+    "correction": "MAINTAIN"
+  },
+  "dish": {
+    "receiver_band": "1"
+  },
+  "csp": {
+    "interface": "https://schema.skao.int/ska-csp-configurescan/4.0",
+    "common": {
+      "config_id": "sbi-mvp01-20200325-00001-science_A",
+      "frequency_band": "1"
+    },
+    "midcbf": {
+        "frequency_band_offset_stream1": 80,
+        "frequency_band_offset_stream2": 80,
+        "correlation": {
+             "processing_regions": [
+                {
+                    "fsp_ids": [1, 2, 3, 4],
+                    "receptors": ["SKA063", "SKA001", "SKA100"],
+                    "start_freq": 350000000,
+                    "channel_width": 13440,
+                    "channel_count": 52080,
+                    "sdp_start_channel_id": 0,
+                    "integration_factor": 1
+                }, {
+                    "fsp_ids": [1, 2, 3, 4],
+                    "receptors": ["SKA063", "SKA001", "SKA100"],
+                    "start_freq": 548437600,
+                    "channel_width": 13440,
+                    "channel_count": 14880,
+                    "sdp_start_channel_id": 1,
+                    "integration_factor": 10
+                }
+              ]
+            },
+      "vlbi": {}
+    }
+  },
+  "sdp": {
+    "interface": "https://schema.skao.int/ska-sdp-configure/0.4",
+    "scan_type": "science_A"
+  },
+  "tmc": {
+    "scan_duration": 10.0
+  }
+}"""
+
+
 INVALID_MID_CONFIGURE_JSON = """
 {
   "interface": "https://schema.skao.int/ska-tmc-configure/2.1",
@@ -1105,7 +1167,7 @@ INVALID_MID_CONFIGURE_JSON = """
     "receiver_band": "1"
   },
   "csp": {
-    "interface": "https://schema.skao.int/ska-csp-configure/2.0",
+    "interface": "https://schema.skao.int/ska-csp-configurescan/2.0",
     "subarray": {
       "subarray_name": "science period 23"
     },
@@ -1174,10 +1236,6 @@ INVALID_MID_CONFIGURE_JSON = """
           "zoom_window_tuning": 650000
         }
       ],
-      "vlbi": {}
-    },
-    "pss": {},
-    "pst": {}
   },
   "sdp": {
     "interface": "https://schema.skao.int/ska-sdp-configure/0.4",
@@ -1188,7 +1246,7 @@ INVALID_MID_CONFIGURE_JSON = """
   }
 }"""
 
-VALID_MID_CONFIGURE_OBJECT = ConfigureRequest(
+VALID_MID_CONFIGURE_OBJECT_2_3 = ConfigureRequest(
     interface="https://schema.skao.int/ska-tmc-configure/2.3",
     transaction_id="txn-....-00001",
     pointing=PointingConfiguration(
@@ -1206,7 +1264,7 @@ VALID_MID_CONFIGURE_OBJECT = ConfigureRequest(
         scan_type="science_A",
     ),
     csp=CSPConfiguration(
-        interface="https://schema.skao.int/ska-csp-configure/2.0",
+        interface="https://schema.skao.int/ska-csp-configurescan/2.0",
         subarray=SubarrayConfiguration(subarray_name="science period 23"),
         common=CommonConfiguration(
             config_id="sbi-mvp01-20200325-00001-science_A",
@@ -1239,6 +1297,60 @@ VALID_MID_CONFIGURE_OBJECT = ConfigureRequest(
                     zoom_window_tuning=650000,
                 ),
             ],
+            vlbi_config={},
+        ),
+    ),
+    tmc=TMCConfiguration(scan_duration=timedelta(seconds=10)),
+)
+
+VALID_MID_CONFIGURE_OBJECT_4_0 = ConfigureRequest(
+    interface="https://schema.skao.int/ska-tmc-configure/4.0",
+    transaction_id="txn-....-00001",
+    pointing=PointingConfiguration(
+        target=Target(
+            ra="21:08:47.92",
+            dec="-88:57:22.9",
+            target_name="Polaris Australis",
+            reference_frame="icrs",
+        ),
+        correction=PointingCorrection.MAINTAIN,
+    ),
+    dish=DishConfiguration(receiver_band=ReceiverBand.BAND_1),
+    sdp=SDPConfiguration(
+        interface="https://schema.skao.int/ska-sdp-configure/0.4",
+        scan_type="science_A",
+    ),
+    csp=CSPConfiguration(
+        interface="https://schema.skao.int/ska-csp-configurescan/4.0",
+        common=CommonConfiguration(
+            config_id="sbi-mvp01-20200325-00001-science_A",
+            frequency_band=ReceiverBand.BAND_1,
+        ),
+        midcbf=MidCBFConfiguration(
+            frequency_band_offset_stream1=80,
+            frequency_band_offset_stream2=80,
+            correlation=CorrelationConfiguration(
+                processing_regions=[
+                    ProcessingRegionConfiguration(
+                        fsp_ids=[1, 2, 3, 4],
+                        receptors=["SKA063", "SKA001", "SKA100"],
+                        start_freq=350000000,
+                        channel_width=13440,
+                        channel_count=52080,
+                        sdp_start_channel_id=0,
+                        integration_factor=1,
+                    ),
+                    ProcessingRegionConfiguration(
+                        fsp_ids=[1, 2, 3, 4],
+                        receptors=["SKA063", "SKA001", "SKA100"],
+                        start_freq=548437600,
+                        channel_width=13440,
+                        channel_count=14880,
+                        sdp_start_channel_id=1,
+                        integration_factor=10,
+                    ),
+                ]
+            ),
             vlbi_config={},
         ),
     ),
@@ -1436,9 +1548,17 @@ def partial_invalidator(o: ConfigureRequest):
         ),
         (
             ConfigureRequest,
-            VALID_MID_CONFIGURE_OBJECT,
+            VALID_MID_CONFIGURE_OBJECT_2_3,
             mid_invalidator,
-            VALID_MID_CONFIGURE_JSON,
+            VALID_MID_CONFIGURE_JSON_2_3,
+            INVALID_MID_CONFIGURE_JSON,
+            True,
+        ),
+        (
+            ConfigureRequest,
+            VALID_MID_CONFIGURE_OBJECT_4_0,
+            mid_invalidator,
+            VALID_MID_CONFIGURE_JSON_4_0,
             INVALID_MID_CONFIGURE_JSON,
             True,
         ),
@@ -1484,9 +1604,9 @@ def partial_invalidator(o: ConfigureRequest):
         ),
         (
             ConfigureRequest,
-            VALID_MID_CONFIGURE_OBJECT,
+            VALID_MID_CONFIGURE_OBJECT_2_3,
             None,
-            VALID_MID_CONFIGURE_JSON,
+            VALID_MID_CONFIGURE_JSON_2_3,
             None,
             True,
         ),