+++ CHANGELOG.md
@@ -4,11 +4,18 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+Unreleased Changes
+******************
+* Data model changes to support ADR-99:
+  - Deprecated csp.FSPConfiguration
+  - Deprecated csp.CBFConfiguration
+  - Adds new MidCBFConfiguration.
+
 12.1.0
 **********
 * Integrated latest OSD version into CDM.
-* After integration of OSD due to newly added semantic validation rules 
-  few assign and scan testcases were failing so fixed those. 
+* After integration of OSD due to newly added semantic validation rules
+  few assign and scan testcases were failing so fixed those.
 
 12.0.0
 **********